requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
PyJWT==2.10.1
python-dotenv==1.0.0
marshmallow==3.20.1
requests==2.32.5
yfinance==0.2.48
